change cpuload to be quot/remainder, remove openssl signing space

+++ linkdata/lh_api5plugin/LH_Qt_QFileInfo.h
@@ -56,7 +56,7 @@ public:
 
     virtual void setup_resize( size_t needed )
     {
-        array_.resize(needed);
+        array_.resize((int) needed);
         item_.param.size = array_.capacity();
         item_.data.s = array_.data();
         return;
